+++ src/main/java/com/imsweb/validation/runtime/ParsedLookups.java
@@ -3,9 +3,20 @@
  */
 package com.imsweb.validation.runtime;
 
+/**
+ * Interface used for pre-parsed referenced lookups.
+ */
 public interface ParsedLookups {
 
+    /**
+     * Returns the validator ID
+     * @return validator ID, never null
+     */
     String getValidatorId();
 
+    /**
+     * Returns the validator version
+     * @return validator version, never null
+     */
     String getValidatorVersion();
 }