+++ src/main/java/com/imsweb/validation/runtime/ParsedProperties.java
@@ -3,9 +3,20 @@
  */
 package com.imsweb.validation.runtime;
 
+/**
+ * Interface used for pre-parsed properties.
+ */
 public interface ParsedProperties {
 
+    /**
+     * Returns the validator ID
+     * @return validator ID, never null
+     */
     String getValidatorId();
 
+    /**
+     * Returns the validator version
+     * @return validator version, never null
+     */
     String getValidatorVersion();
 }